+++ src/f5_tts/model/backbones/unett.py
@@ -54,7 +54,7 @@ class TextEmbedding(nn.Module):
         text = text + 1  # use 0 as filler token. preprocess of batch pad -1, see list_str_to_idx()
         text = text[:, :seq_len]  # curtail if character tokens are more than the mel spec tokens
         batch, text_len = text.shape[0], text.shape[1]
-        text = F.pad(text, (0, seq_len - text_len), value=0)
+        text = F.pad(text, (0, seq_len - text_len), value=0).long()
         if self.mask_padding:
             text_mask = text == 0
